+++ packages/react/src/components/TextInput/index.tsx
@@ -1,14 +1,16 @@
 import { ComponentProps, ComponentRef, forwardRef } from 'react'
 import { Input, Prefix, TextInputContainer } from './style'
 
-export interface TextInputProps extends ComponentProps<typeof Input> {
+export interface TextInputProps
+  extends Omit<ComponentProps<typeof Input>, 'size'> {
   prefix?: string
+  size?: ComponentProps<typeof TextInputContainer>['size']
 }
 
 export const TextInput = forwardRef<ComponentRef<typeof Input>, TextInputProps>(
-  ({ prefix, ...props }: TextInputProps, ref) => {
+  ({ prefix, size, ...props }: TextInputProps, ref) => {
     return (
-      <TextInputContainer>
+      <TextInputContainer size={size}>
         {!!prefix && <Prefix>{prefix}</Prefix>}
         <Input ref={ref} {...props} />
       </TextInputContainer>