+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-alfaui/react
 
+## 2.1.0
+
+### Minor Changes
+
+- add toast and tooltip components
+
 ## 2.0.1
 
 ### Patch Changes