+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ignite-alfaui/tokens
 
+## 2.2.0
+
+### Minor Changes
+
+- add toast and tooltip components
+
 ## 2.1.0
 
 ### Minor Changes